integration: rename default_policy -> load_balancing
I will put a regression test for a custom LBP later. The file
has only one test, so we can put all LBP related tests there.
