+++ scylla/tests/integration/main.rs
@@ -4,11 +4,11 @@ mod consistency;
 mod cql_collections;
 mod cql_types;
 mod cql_value;
-mod default_policy;
 mod execution_profiles;
 mod history;
 mod hygiene;
 mod large_batch_statements;
+mod load_balancing;
 mod lwt_optimisation;
 mod metadata_custom_timeouts;
 mod new_session;
